test: fix coverage report for the fee server, refactor
